Bug 3292: Invalid Range exception in setDpnId
1. Change dpnId from long to BigInteger

+++ fibmanager/fibmanager-api/src/main/java/org/opendaylight/fibmanager/api/IFibManager.java
@@ -1,6 +1,8 @@
 package org.opendaylight.fibmanager.api;
 
+import java.math.BigInteger;
+
 public interface IFibManager {
-    void populateFibOnNewDpn(long dpnId, long vpnId, String rd);
-    void cleanUpDpnForVpn(long dpnId, long vpnId, String rd);
+    void populateFibOnNewDpn(BigInteger dpnId, long vpnId, String rd);
+    void cleanUpDpnForVpn(BigInteger dpnId, long vpnId, String rd);
 }
