+++ interfacemgr/interfacemgr-api/src/main/java/org/opendaylight/vpnservice/interfacemgr/interfaces/IInterfaceManager.java
@@ -8,6 +8,8 @@
 
 package org.opendaylight.vpnservice.interfacemgr.interfaces;
 
+import java.math.BigInteger;
+
 import java.util.List;
 import org.opendaylight.vpnservice.mdsalutil.ActionInfo;
 import org.opendaylight.vpnservice.mdsalutil.MatchInfo;
@@ -15,8 +17,8 @@ import org.opendaylight.vpnservice.mdsalutil.MatchInfo;
 public interface IInterfaceManager {
 
     public Long getPortForInterface(String ifName);
-    public long getDpnForInterface(String ifName);
-    public String getEndpointIpForDpn(long dpnId);
+    public BigInteger getDpnForInterface(String ifName);
+    public String getEndpointIpForDpn(BigInteger dpnId);
     public List<MatchInfo> getInterfaceIngressRule(String ifName);
     public List<ActionInfo> getInterfaceEgressActions(String ifName);
