+++ vpnmanager/vpnmanager-api/src/main/java/org/opendaylight/vpnmanager/api/IVpnManager.java
@@ -1,10 +1,11 @@
 package org.opendaylight.vpnmanager.api;
 
-import java.util.Collection;
+import java.math.BigInteger;
 
+import java.util.Collection;
 import org.opendaylight.fibmanager.api.IFibManager;
 
 public interface IVpnManager {
-    Collection<Long> getDpnsForVpn(long vpnId);
+    Collection<BigInteger> getDpnsForVpn(long vpnId);
     void setFibService(IFibManager fibManager);
 }
